% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_species}
\alias{all_species}
\title{World Health Organization TB data}
\format{
## `all_species`
A vector of strings
}
\source{
obtained from SwissProt
}
\usage{
all_species
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
