% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrDIA.R
\name{corrDIA}
\alias{corrDIA}
\title{corrDIA}
\usage{
corrDIA(
  data,
  transformation = c("none", "log2"),
  tit = "",
  data_type = c("intensity", "Top3", "iBAQ", "all"),
  gradient_color = c("#09009D", "#ffffff", "#BE0010"),
  plot_pairs = FALSE
)
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2 or none)}

\item{tit}{Title of your plot}

\item{data_type}{The type of data you want to visualize; either 'intensity', 'Top3', 'iBAQ' or 'all'.}

\item{gradient_color}{Three colors for the printed gradient color on the correlation plot}

\item{plot_pairs}{Plot pairwise correlation plot. If FALSE, will plot correlation matrix plot.}
}
\value{
ggplot2 correlation plot
}
\description{
Plot correlation plot from each fraction from processed data from DIAnn
}
