% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityDIA.R
\name{densityDIA}
\alias{densityDIA}
\title{densityDIA}
\usage{
densityDIA(
  data,
  transformation = c("none", "log2"),
  area = FALSE,
  tit = "",
  data_type = c("intensity", "Top3", "iBAQ", "all"),
  colorspace = NULL
)
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2 or none)}

\item{area}{logical; print the area under curve}

\item{tit}{Title of your plot}

\item{data_type}{The type of data you want to visualize; either 'intensity', 'Top3', 'iBAQ' or 'all'.}

\item{colorspace}{A vector of colors of same length as the number of condition you have to plot.
If NULL, will plot default ggplot2 colors.}
}
\value{
ggplot2 density graph
}
\description{
Print density from each fraction from processed data from DIAnn
}
