% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_matrix.R
\name{diann_matrix}
\alias{diann_matrix}
\title{diann_matrix}
\usage{
diann_matrix(
  x,
  id.header = "Precursor.Id",
  quantity.header = "Precursor.Normalised",
  proteotypic.only = FALSE,
  q = 0.01,
  protein.q = 1,
  pg.q = 0.01,
  gg.q = 1,
  get_pep = FALSE,
  only_pepall = FALSE,
  margin = -10,
  Top3 = FALSE,
  method = c("max", "sum")
)
}
\arguments{
\item{x}{data, output from diann_load}

\item{id.header}{Id column name (protein, gene, ...)}

\item{quantity.header}{Quantity column name}

\item{proteotypic.only}{logical; Only proteotypic peptides and the
respective proteins should be considered?}

\item{q}{Precursor q-value threshold}

\item{protein.q}{Uniquely identified protein q-value threshold}

\item{pg.q}{Protein group q-value threshold}

\item{gg.q}{Gene group q-value threshold}

\item{get_pep}{logical; get peptide count?}

\item{only_pepall}{logical; should only keep peptide counts all or also
peptide counts for each fractions?}

\item{margin}{quantities below exp(margin) might be treated as NA}

\item{Top3}{logical; get Top3 absolute quantification}

\item{method}{When for one identifier there are several values, take either
maximum of these or sum them all. When using 'sum', any additional
information than the id.header should be ignored. 'max' reports for only
one id, 'sum' get the sum of the intensities of all ids.}
}
\value{
A dataframe containing the quantities from the id you selected
}
\description{
Generate matrix with quantities (zero will be replaced by NA values) and
eventually peptides count.
}
