% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diann_maxlfq.R
\name{diann_maxlfq}
\alias{diann_maxlfq}
\title{diann_maxlfq}
\usage{
diann_maxlfq(
  x,
  sample.header = "File.Name",
  group.header = "Protein.Names",
  id.header = "Precursor.Id",
  quantity.header = "Precursor.Normalised",
  margin = -10,
  count_pep = TRUE,
  only_countsall = FALSE,
  Top3 = FALSE
)
}
\arguments{
\item{x}{data, output from diann_load}

\item{sample.header}{Sample id column name}

\item{group.header}{Colulmn name corresponding to the group Id, e.g. protein Id}

\item{id.header}{Id column name (protein, gene, ...)}

\item{quantity.header}{Quantity column name}

\item{margin}{quantities below exp(margin) might be treated as NA}

\item{count_pep}{logical; get peptide count?}

\item{only_countsall}{logical; should only keep peptide counts all or also peptide counts for each fractions?}

\item{Top3}{logical; get Top3 absolute quantification}
}
\value{
A dataframe containing the quantities from the id you selected
}
\description{
Quantify using the MaxLFQ algorithm https://doi.org/10.1074/mcp.M113.031591.
This function can be used to calculate protein quantities from peptide/precursor/fragment quantities or, for example,
to calculate peptide quantities from precursor/fragment quantities.
Totally based on function from diann-rpackage from Vadim Demichev
}
