% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bestwind.R
\name{get_bestwind}
\alias{get_bestwind}
\title{get_bestwind}
\usage{
get_bestwind(data, bins = 25, per_frac = FALSE)
}
\arguments{
\item{data}{The report-lib data from DIA-nn which contains the \emph{ProductMz} column.
It can be a path to the file or the dataframe corresponding to the file
(need to have \emph{FileName} and \emph{ProductMz} columns in that case).}

\item{bins}{The number of windows you want to have.}

\item{per_frac}{If \code{FALSE}, will select the best windows from all fraction without differentiation.
Else, it will select the best window for each fraction and then do the mean of those.}
}
\value{
A list containing a data frame with \emph{FileName}, \emph{ProductMz} and a \emph{bins} column;
        four plots corresponding to bar plots / histogramm from the distribution of the \emph{ProductMz}
        with and without the window selection.
}
\description{
Get the best m/z windows for DIA according a number of window and the report-lib from DIA-nn.
}
\details{
In the end, the distribution of the \emph{ProductMz} per window will be uniform.
}
