% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getallseq.R
\name{getallseq}
\alias{getallseq}
\title{getallseq}
\usage{
getallseq(
  spec = "Saccharomyces cerevisiae",
  pr_id = "P09938",
  bank_name = "swissprot",
  fasta_file = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{spec}{The species name corresponding to the proteins from \code{pr_id}.}

\item{pr_id}{The proteins that you want to get the theoretical peptide sequence from.}

\item{bank_name}{The name of the data bank you want to use to get theoretical peptide.
If you want to use you're own fasta files, you can put the path (or several in a double)
to your fasta files and set fasta_file to TRUE.}

\item{fasta_file}{Logical to tell if you use fasta file in \code{bank_name}.}

\item{verbose}{Logical to tell if you want to print advancement.}
}
\value{
A list containing the peptide sequence from each protein.
}
\description{
Get theoretical peptide sequence from proteins (from data bank or fasta files), thanks to \pkg{seqinr} package.
}
\details{
Charif D, Lobry J (2007). “SeqinR 1.0-2: a contributed package to the R project for statistical computing devoted to
biological sequences retrieval and analysis.” In Bastolla U, Porto M, Roman H, Vendruscolo M (eds.),
Structural approaches to sequence evolution: Molecules, networks, populations, series Biological and Medical Physics,
Biomedical Engineering, 207-232. Springer Verlag, New York. ISBN : 978-3-540-35305-8.

If the protein is grouped, it will only take the first protein.
}
