% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iBAQ.R
\name{getProtease}
\alias{getProtease}
\title{getProtease}
\usage{
getProtease(protease = "trypsin")
}
\arguments{
\item{protease}{name of the enzyme; currently "trypsin" or "lys-c"}
}
\value{
a regular expression
}
\description{
Returns a regular corresponding to the given enzyme in order to digest a
proteic sequence.
}
\note{
function from SafeQuant R package by Erik Ahrne (2021). SafeQuant: A
  Toolbox for the Analysis of Proteomics Data. R package version 2.3.4.
}
\examples{
getProtease()
}
