% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDIA.R
\name{heatmapDIA}
\alias{heatmapDIA}
\title{heatmapDIA}
\usage{
heatmapDIA(
  data,
  transformation = c("log2", "z.score_proteins", "z.score_fraction", "none"),
  maxna = 0,
  print_val = TRUE,
  nm_id = NULL
)
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2, z.score on proteins or on fractions, none)}

\item{maxna}{How many NAs do you authorize per proteins}

\item{print_val}{logical; do you want to print values on heatmpa ?}

\item{nm_id}{The name of column that contains the IDs (if NULL, will take it automatically)}
}
\value{
Plotly heatmap
}
\description{
Print interactive heatmap of processed data from DIAnn
}
