% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationDIA.R
\name{imputationDIA}
\alias{imputationDIA}
\title{imputationDIA}
\usage{
imputationDIA(
  data,
  transformation = c("none", "log2"),
  iteration = 3,
  method = c("norm.predict", "zeros", "pmm", "midastouch", "sample", "cart", "rf",
    "mean", "norm", "norm.nob", "norm.boot", "lasso.norm", "lasso.select.norm", "knn",
    "MinDet", "MinProb", "QRILC")
)
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2 or none)}

\item{iteration}{Number of iterations for the imputation. See \code{\link[mice]{mice}} and \code{\link[imputeLCMD]{imputeLCMD}} for more details.}

\item{method}{Method used for the imputation. See \code{\link[mice]{mice}} and \code{\link[imputeLCMD]{imputeLCMD}} for more details.}
}
\value{
The imputed data
}
\description{
Perform data imputation for your DIA data
}
\details{
If your data contains iBAQ and or Top3 absolute quantification, then their column
  names should start with 'iBAQ_' and 'Top3_' respectively.
  Other numeric columns like the number of peptides used for quantification or the
  number of theoritical peptides will not be imputed, please keep the default name obtained
  with the app or the functions \code{diann_matrix}  or \code{diann_maxlfq}.
}
