% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_process.R
\name{report_process}
\alias{report_process}
\title{report_process}
\usage{
report_process(
  data,
  header.id = "Protein.Group",
  sample.id = "File.Name",
  quantity.id = "Precursor.Normalised",
  secondary.id = "Precursor.Id",
  id_to_add = c("Protein.Names", "First.Protein.Description", "Genes"),
  qv = 0.01,
  pg.qv = 0.01,
  p.qv = 1,
  gg.qv = 1,
  only_proteotypic = TRUE,
  get_pep = TRUE,
  only_pepall = FALSE,
  get_Top3 = FALSE,
  get_iBAQ = FALSE,
  fasta = NULL,
  species = NULL,
  peptide_length = c(5, 36),
  format = c("xlsx", "csv", "txt")
)
}
\arguments{
\item{data}{Path to data (report from DIAnn)}

\item{header.id}{Id column name (protein, gene, ...)}

\item{sample.id}{Id column that contains the fractions names}

\item{quantity.id}{Quantity column name (Precursor.Normalised)}

\item{secondary.id}{Id column that contains precursor (usefull when get_pep is TRUE)}

\item{id_to_add}{Some secondary id you want to keep in your final data}

\item{qv}{Precursor q-value threshold}

\item{pg.qv}{Protein group q-value threshold}

\item{p.qv}{Uniquely identified protein q-value threshold}

\item{gg.qv}{Gene group q-value threshold}

\item{only_proteotypic}{If TRUE, will only keep proteotypic}

\item{get_pep}{logical; get peptide count ?}

\item{only_pepall}{logical; should only keep peptide counts all or also peptide counts for each fractions ?}

\item{get_Top3}{If TRUE, will calculate Top3 quantification
(only use when \code{header.id} is set to 'Protein.group' or 'Genes)}

\item{get_iBAQ}{If TRUE, will calculate iBAQ quantification
(only use when \code{header.id} is set to 'Protein.group' )}

\item{fasta}{When iBAQ is set to TRUE, you can add path to fasta files.
If NULL, will get theoretical peptides from swissprot bank (quite long).}

\item{species}{When fasta is NULL, you need to specify the specie from your data}

\item{peptide_length}{The minimum and maximum peptide length}

\item{format}{Format from your saved file (only xlsx, csv and txt are supported)}
}
\description{
Process data from DIAnn in a single workflow, save it and also save some quality check plots.
}
