% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_report}
\alias{small_report}
\title{World Health Organization TB data}
\format{
## `small_report`
A data frame with 2992 rows and 40 columns:
\describe{
  \item{File.Name}{File.Name}
  \item{Protein.Group}{Protein.Group}
  \item{Protein.Ids}{Protein.Ids}
  \item{Protein.Names}{Protein.Names}
  \item{Genes}{Genes}
  \item{PG.Quantity}{PG.Quantity}
  \item{PG.Normalised}{PG.Normalised}
  \item{Genes.Quantity}{Genes.Quantity}
  \item{Genes.Normalised}{Genes.Normalised}
  \item{Genes.MaxLFQ}{Genes.MaxLFQ}
  \item{Genes.MaxLFQ.Unique}{Genes.MaxLFQ.Unique}
  \item{Modified.Sequence}{Modified.Sequence}
  \item{Stripped.Sequence}{Stripped.Sequence}
  \item{Precursor.Id}{Precursor.Id}
  \item{Precursor.Charge}{Precursor.Charge}
  \item{Q.Value}{Q.Value}
  \item{Protein.Q.Value}{Protein.Q.Value}
  \item{PG.Q.Value}{PG.Q.Value}
  \item{GG.Q.Value}{GG.Q.Value}
  \item{Proteotypic}{Proteotypic}
  \item{Precursor.Quantity}{Precursor.Quantity}
  \item{Precursor.Normalised}{Precursor.Normalised}
  \item{Label.Ratio}{Label.Ratio}
  \item{RT}{RT}
  \item{RT.Start}{RT.Start}
  \item{RT.Stop}{RT.Stop}
  \item{iRT}{iRT}
  \item{Predicted.RT}{Predicted.RT}
  \item{Predicted.iRT}{Predicted.iRT}
  \item{First.Protein.Description}{First.Protein.Description}
  \item{Lib.Q.Value}{Lib.Q.Value}
  \item{Ms1.Profile.Corr}{Ms1.Profile.Corr}
  \item{Ms1.Area}{Ms1.Area}
  \item{Evidence}{Evidence}
  \item{CScore}{CScore}
  \item{Decoy.Evidence}{Decoy.Evidence}
  \item{Decoy.CScore}{Decoy.CScore}
  \item{Fragment.Quant.Raw}{Fragment.Quant.Raw}
  \item{Fragment.Quant.Corrected}{Fragment.Quant.Corrected}
  \item{Fragment.Correlations}{Fragment.Correlations}
}
}
\source{
this data were obtained thanks to diann-rpackage from Vadim Demichev
}
\usage{
small_report
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
