% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validDIA.R
\name{validDIA}
\alias{validDIA}
\title{validDIA}
\usage{
validDIA(
  data,
  transformation = c("none", "log2"),
  tit = "",
  data_type = c("intensity", "Top3", "iBAQ", "all"),
  design = NULL,
  to_check = NULL,
  prop_cut = 0.75
)
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2 or none)}

\item{tit}{Title of your plot}

\item{data_type}{The type of data you want to visualize; either 'intensity', 'Top3', 'iBAQ' or 'all'.}

\item{design}{The design of your experiment that match your columns names.
For example, if you have columns named '37_B1_Treatment', put \code{c("temperature", "replicate", "condition")}.
If NULL, will take all values.}

\item{to_check}{The grouping variable (same as in your design; like 'replicate' for example)}

\item{prop_cut}{The minimum proportion to show on the graph}
}
\value{
ggplot2 graph showing proportion of non-missing values
}
\description{
Print plot showing proportion of non-missing values for each protein (or else) among the experiment
}
\details{
If you have replicates or several conditions and you want to see the proportion of non-missing values
         for each replicate for example. Then, the id column in your data must be the first one;
         as in the output of the shiny app of DIAgui.
}
