% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityDIA.R
\name{densityDIA}
\alias{densityDIA}
\title{densityDIA}
\usage{
densityDIA(data, transformation = c("log2", "none"), area = FALSE, tit = "")
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2 or none)}

\item{area}{logical; print the area under curve}

\item{tit}{Title of your plot}
}
\value{
ggplot2 density graph
}
\description{
Print density from each fraction from processed data from DIAnn
}
