% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapDIA.R
\name{heatmapDIA}
\alias{heatmapDIA}
\title{heatmapDIA}
\usage{
heatmapDIA(
  data,
  transformation = c("none", "log2", "z.score_proteins", "z.score_fraction"),
  maxna = 0,
  print_val = TRUE,
  nm_id = NULL,
  data_type = c("intensity", "Top3", "iBAQ", "all"),
  gradient_color = c("#09009D", "#ffffff", "#BE0010"),
  static = FALSE
)
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2, z.score on proteins or on fractions, none)}

\item{maxna}{How many NAs do you authorize per proteins}

\item{print_val}{logical; do you want to print values on heatmpa ?}

\item{nm_id}{The name of column that contains the IDs (if NULL, will take it automatically)}

\item{data_type}{The type of data you want to visualize; either 'intensity', 'Top3', 'iBAQ' or 'all'.}

\item{gradient_color}{Three colors for the printed gradient color on the heatmap}

\item{static}{Logical to return a static plot (ggplot object) or an interactive one (plotly object)}
}
\value{
Plotly heatmap
}
\description{
Print interactive heatmap of processed data from DIAnn
}
