% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDS_DIA.R
\name{MDS_DIA}
\alias{MDS_DIA}
\title{MDS_DIA}
\usage{
MDS_DIA(
  data,
  transformation = c("log2", "none"),
  tit = "",
  data_type = c("intensity", "Top3", "iBAQ", "all")
)
}
\arguments{
\item{data}{Processed data from DIAnn (from iq processing or diann_matrix for example)}

\item{transformation}{Which transformation do you want to apply (log2 or none)}

\item{tit}{Title of your plot}

\item{data_type}{The type of data you want to visualize; either 'intensity', 'Top3', 'iBAQ' or 'all'.}
}
\value{
ggplot2 MDS graph
}
\description{
Print MDS of processed data from DIAnn
}
